[PLAYER-2375] Rename ChromecastV3SampleApp to ChromecastSampleApp
