+++ ChromecastSampleApp/app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion '26.0.2'
 
     defaultConfig {
-        applicationId "com.ooyala.sample.ChromecastSampleV3App"
+        applicationId "com.ooyala.sample.ChromecastSampleApp"
         minSdkVersion 19
         targetSdkVersion 26
         versionCode 1
