+++ ChromecastSampleApp/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.ooyala.chromecastv3sampleapp">
+    package="com.ooyala.sample">
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
@@ -26,7 +26,7 @@
              android:launchMode="singleTop"/>
         <meta-data
             android:name="com.google.android.gms.cast.framework.OPTIONS_PROVIDER_CLASS_NAME"
-            android:value="com.ooyala.chromecastv3sampleapp.CastOptionsProvider" />
+            android:value="com.ooyala.sample.CastOptionsProvider" />
 
     </application>
