+++ ChromecastSampleApp/app/src/main/res/layout/cast_video_view.xml
