+++ ChromecastSampleApp/app/src/main/res/layout/list_activity_layout.xml
