+++ ChromecastSampleApp/app/src/main/res/layout/list_activity_list_item.xml
