+++ ChromecastSampleApp/app/src/main/res/layout/player_activity.xml
