+++ ChromecastSampleApp/app/src/main/res/layout/start_view.xml
