+++ ChromecastSampleApp/app/src/main/res/mipmap-xxxhdpi/ic_launcher.png
