+++ ChromecastSampleApp/app/src/main/res/values/colors.xml
