+++ ChromecastSampleApp/gradle.properties
