+++ ChromecastSampleApp/settings.gradle
